+++ src/datagrid/datagrid.reducer.js
@@ -76,13 +76,17 @@ export default function datagridReducer(state = INITIAL_STATE, action) {
         ], List(), items => items.push(Immutable.fromJS(action.columnDefaultValues)));
 
     case TYPES.PLATFORM_DATAGRID_REMOVE_ITEM: {
-      const itemId = state.getIn([action.id, 'data', action.index, ...action.idKeyPath]);
       const allDataIndex = state.getIn([action.id, 'allData'], List())
-        .findIndex(item => (item.getIn(action.idKeyPath) === itemId));
+        .findIndex(item => (item.getIn(action.idKeyPath) === action.rowId));
+      const dataIndex = state.getIn([action.id, 'data'], List())
+        .findIndex(item => (item.getIn(action.idKeyPath) === action.rowId));
       return state
-        .deleteIn([action.id, 'data', action.index])
+        .deleteIn([action.id, 'data', dataIndex])
         .deleteIn([action.id, 'allData', allDataIndex])
-        .deleteIn([action.id, 'editData', itemId]);
+        .deleteIn([action.id, 'editData', action.rowId])
+        .deleteIn([action.id, 'cellMessages', 'error', action.rowId])
+        .deleteIn([action.id, 'cellMessages', 'info', action.rowId])
+        .deleteIn([action.id, 'cellMessages', 'warning', action.rowId]);
     }
 
     case TYPES.PLATFORM_DATAGRID_REMOVE_NEW_ITEM:
@@ -113,7 +117,9 @@ export default function datagridReducer(state = INITIAL_STATE, action) {
 
     case TYPES.PLATFORM_DATAGRID_EXTEND_DATA: {
       const allData = state.getIn([action.id, 'allData']);
-      const extendedData = allData.concat(action.data);
+      const extendedData = action.prepend ?
+        action.data.concat(allData) : allData.concat(action.data);
+
       return state
         .setIn([action.id, 'data'], extendedData)
         .setIn([action.id, 'allData'], extendedData);